fix: dont call hook inside useMemo()
